+++ packages/windows_foundation/lib/src/internal/extensions/async.dart
