+++ packages/windows_foundation/lib/src/internal/extensions/bool.dart
