+++ packages/windows_foundation/lib/src/internal/extensions/comobject.dart
