+++ packages/windows_foundation/lib/src/internal/extensions/datetime.dart
