+++ packages/windows_foundation/lib/src/internal/extensions/double.dart
