+++ packages/windows_foundation/lib/src/internal/extensions/duration.dart
