+++ packages/windows_foundation/lib/src/internal/extensions/enum.dart
