+++ packages/windows_foundation/lib/src/internal/extensions/extensions.dart
@@ -2,18 +2,18 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-export 'async_helpers.dart';
-export 'bool_conversions.dart';
-export 'comobject_helper.dart';
-export 'datetime_conversions.dart';
-export 'double_conversions.dart';
-export 'duration_conversions.dart';
-export 'enum_conversions.dart';
-export 'guid_conversions.dart';
-export 'iinspectable_helpers.dart';
-export 'int_conversions.dart';
-export 'ipropertyvalue_helpers.dart';
-export 'object_conversions.dart';
-export 'string_conversions.dart';
-export 'struct_conversions.dart';
-export 'uri_conversions.dart';
+export 'async.dart';
+export 'bool.dart';
+export 'comobject.dart';
+export 'datetime.dart';
+export 'double.dart';
+export 'duration.dart';
+export 'enum.dart';
+export 'guid.dart';
+export 'iinspectable.dart';
+export 'int.dart';
+export 'ipropertyvalue.dart';
+export 'object.dart';
+export 'string.dart';
+export 'struct.dart';
+export 'uri.dart';
