+++ packages/windows_foundation/lib/src/internal/extensions/guid.dart
