+++ packages/windows_foundation/lib/src/internal/extensions/iinspectable.dart
