+++ packages/windows_foundation/lib/src/internal/extensions/int.dart
