+++ packages/windows_foundation/lib/src/internal/extensions/string.dart
