+++ packages/windows_foundation/lib/src/internal/extensions/struct.dart
