+++ packages/windows_foundation/lib/src/internal/extensions/uri.dart
