+++ packages/windows_foundation/test/internal/extensions/bool_test.dart
