+++ packages/windows_foundation/test/internal/extensions/double_test.dart
