+++ packages/windows_foundation/test/internal/extensions/duration_test.dart
