+++ packages/windows_foundation/test/internal/extensions/enum_test.dart
