+++ packages/windows_foundation/test/internal/extensions/guid_test.dart
