+++ packages/windows_foundation/test/internal/extensions/int_test.dart
