+++ packages/windows_foundation/test/internal/extensions/object_test.dart
