+++ packages/windows_foundation/test/internal/extensions/struct_test.dart
