+++ packages/windows_foundation/test/internal/extensions/uri_test.dart
