+++ packages/windows_foundation/test/struct_test.dart
@@ -2,8 +2,6 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-@TestOn('windows')
-
 import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
