+++ packages/winrtgen/lib/src/models/models.dart
@@ -6,8 +6,8 @@ export 'array_passing_style.dart';
 export 'asset_manager.dart';
 export 'generic_type.dart';
 export 'json.dart';
-export 'metadata_store.dart';
 export 'projection_kind.dart';
 export 'type_analyzer.dart';
 export 'type_arg_kind.dart';
 export 'winrt_docs_service.dart';
+export 'winrt_metadata_store.dart';
