+++ packages/winrtgen/lib/src/models/winrt_metadata_store.dart
